+++ src/dev/testClass/db/TestDBUpdate01.ts
@@ -6,7 +6,7 @@ export default class TestDBUpdate01 extends CoreDBUpdate<any> {
   }
 
   async performe(): Promise<boolean> {
-    this.getDb().log(`${this.srcVersion}->${this.tarVersion}`);
+    this.getDb().debug(`${this.srcVersion}->${this.tarVersion}`);
     return true;
   }
 }