+++ src/modules/index.ts
@@ -1,7 +1,7 @@
-import DefaultLogger from './log/DefaultLogger';
+import DefaultLogger, { DC } from './log/DefaultLogger';
 import InMemDB from './db/InMemDB';
 import InMemCache from './cache/InMemCache';
 
-export { DefaultLogger, InMemDB, InMemCache };
+export { DefaultLogger, InMemDB, InMemCache, DC };
 export * from './crypto';
 export * from './env';